==> notebooks/2.0-jkg-data-wrangling.ipynb <==
# Data Wrangling for Capstone Project 1

This notebook begins with importing the raw dataset for my capstone project and moves through some basic cleaning and reorganization to produce a pandas DataFrame ready for further analysis. Future exploration will undoubtedly reveal additional data wrangling steps that need to be taken, but for now the dataset produced by this notebook is clean enough that it can be investigated in much more detail than in its raw form. The overall steps taken in this notebook are as follows:

1. Load data into pandas DataFrame
2. Rename columns
3. Convert data types (dates)
4. Set AppointmentID as index
5. Copy to new, clean DataFrame
6. Create Patients and Appointments DataFrames

**Findings from preliminary exploration:**

*Note:* A preliminary exploration of the dataset that led to these data cleaning steps can be found in [this](./1.0-jkg-initial-data-exploration.ipynb) notebook.

- All fields appear to have no null values.
- PatientId is type float64, but AppointmentID is type int64
- Date(time?) fields: (need to convert)
    - ScheduledDay
    - AppointmentDay
- Rename:
    - 'PatientId' --> 'Patient_ID'
    - 'AppointmentID' --> 'Appointment_ID'
    - 'ScheduledDay' --> 'Scheduled_Date'
    - 'AppointmentDay' --> 'Appointment_Date'
    - 'Neighbourhood' --> 'Neighborhood'
    - 'Scholarship' --> 'Welfare' (see note in reference https://en.wikipedia.org/wiki/Bolsa_Fam%C3%ADlia)
    - 'Hipertension' --> 'Hypertension'
    - 'Handcap' --> 'Disability'
    - 'SMS_received' --> 'SMS_sent'
    - 'No-show' --> 'No_show'

# Package imports
import pandas as pd
import matplotlib.pyplot as plt

## 1. Load Data into pandas DataFrame

# import raw data into DataFrame
df_import = pd.read_csv('../data/raw/KaggleV2-May-2016.csv')

## 2. Rename Columns

# Examine columns
df_import.columns

# Rename columns
df_import.columns = ['Patient_ID',
                  'Appointment_ID',
                  'Gender',
                  'Scheduled_Date',
                  'Appointment_Date',
                  'Age',
                  'Neighborhood',
                  'Welfare',
                  'Hypertension',
                  'Diabetes',
                  'Alcoholism',
                  'Disability',
                  'SMS_sent',
                  'No_show']

## 3. Convert Data Types (dates)

# Examine Data Types
df_import.info()

# Convert Scheduled_Date to DateTime
df_import['Scheduled_Date'] = pd.to_datetime(df_import['Scheduled_Date'])

# Convert Appointment_Date to DateTime
df_import['Appointment_Date'] = pd.to_datetime(df_import['Appointment_Date'])

## 4. Set Appointment_ID as Index

# Set index to Appointment_ID (unique)
df_import.set_index('Appointment_ID')

## 5. Copy to New DataFrame

# Copy df_import to df_clean
df_clean = df_import.copy()

df_clean.info()

## 6. Create Patients and Appointments DataFrames

**Make Patients DataFrame**

df_patients = df_clean.copy()

# Drop unnecessary columns from patients dataframe
# Note: removing 'Age' here because it changes for some patients over time
df_patients.drop(columns=['Scheduled_Date', 'Appointment_Date', 'SMS_sent', 'Age'], inplace=True)

# Convert no_show column to 1/0 for easier calculation
df_patients.No_show.replace(('Yes', 'No'), (1, 0), inplace=True)

# Calculate total number of appointments and number of no_show appointments
patient_appointments = df_patients.groupby('Patient_ID')['Appointment_ID'].count()
patient_noshows = df_patients.groupby('Patient_ID')['No_show'].sum()

# Remove appointment columns (no longer needed after sum/count in previous step)
df_patients.drop(columns=['Appointment_ID', 'No_show'], inplace=True)

# Make DataFrame 'per-patient' by removing duplicates and setting Patient_ID as index
df_patients.drop_duplicates(inplace=True)
df_patients.set_index('Patient_ID', inplace=True)

# Add calculated series to dataframe as columns
patient_appointments.name = 'total_appointments'
df_patients = df_patients.join(patient_appointments)

patient_noshows.name = 'noshow_appointments'
df_patients = df_patients.join(patient_noshows)

# sort completed dataframe
df_patients.sort_index(inplace=True)

**Make Appointments DataFrame**

df_appointments = df_clean.copy()

df_appointments.set_index('Appointment_ID', inplace=True)

# Drop extra columns
df_appointments.drop(columns=['Age', 'Neighborhood', 'Welfare', 'Hypertension', 'Diabetes', 'Alcoholism', 'Disability'], inplace=True)

# Calculate Date Difference (Appointment - Scheduled)
# Convert scheduled date to date since appointment date does not have time included
# BUT...need to also use `.dt.date` with Appointment_Date due to type mismatch otherwise...
df_appointments['date_diff'] = df_appointments['Appointment_Date'].dt.date - df_appointments['Scheduled_Date'].dt.date


# Save DataFrames for use in modeling
df_clean.to_pickle('../data/interim/clean_df.pickle')
df_patients.to_pickle('../data/interim/patients_df.pickle')
df_appointments.to_pickle('../data/interim/appointments_df.pickle')
