==> notebooks/4.0-jkg-models.ipynb <==
# Setup

# package imports
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from sklearn.model_selection import train_test_split, GridSearchCV
from sklearn.linear_model import LogisticRegression
from sklearn.metrics import confusion_matrix, classification_report

# import pickle file
df_appointments = pd.read_pickle('../data/interim/appointments_df.pickle')
df_clean = pd.read_pickle('../data/interim/clean_df.pickle')

# set index for dataframe
df_clean.set_index('Appointment_ID', inplace=True)

# add columns from df_clean back into df_appointments
df_model = df_appointments.join(df_clean, rsuffix='_clean')

# remove duplicate columns from model dataframe
drop_columns = ['Patient_ID_clean',
                'Gender_clean',
                'Scheduled_Date_clean',
                'Appointment_Date_clean',
                'SMS_sent_clean',
                'No_show_clean',
                'Neighborhood']
df_model.drop(columns=drop_columns, inplace=True)

# create new column: days_diff to show number of days between scheduled date and appointment date
df_model['days_diff'] = df_model.date_diff.dt.days

# create dummy vars for days of week
df_model = df_model.join(pd.get_dummies(df_model.Appointment_Date.dt.dayofweek))

# clean up column names
df_model.columns = ['Patient_ID',
                    'Gender',
                    'Scheduled_Date',
                    'Appointment_Date',
                    'SMS_sent',
                    'No_show',
                    'date_diff',
                    'Age',
                    'Welfare',
                    'Hypertension',
                    'Diabetes',
                    'Alcoholism',
                    'Disability',
                    'days_diff',
                    'Mon',
                    'Tue',
                    'Wed',
                    'Thu',
                    'Fri',
                    'Sat']

# extraneous columns not needed for modeling
unneeded_columns = ['Patient_ID',
                    'Scheduled_Date',
                    'Appointment_Date',
                    'No_show',
                    'date_diff',
                    'Gender']

# Need to produce arrays for the features and the response variable
y = df_model['No_show'].values
X = df_model.drop(columns=unneeded_columns).values

# rename 'Yes'/'No' to 'No-show'/'Show' for clarity
rename = y.copy()
for key, value in {'Yes': 'No-show', 'No': 'Show'}.items(): rename[y == key] = value
y = rename

# Logistic Regression

# Create training and test sets
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size = 0.4, random_state=42)

# Create the classifier: logreg
logreg = LogisticRegression()

# Fit the classifier to the training data
logreg.fit(X_train, y_train)

# Predict the labels of the test set: y_pred
y_pred = logreg.predict(X_test)

# Compute and print the confusion matrix and classification report
logreg_cm = confusion_matrix(y_test, y_pred)
true_pos, false_neg, false_pos, true_neg = logreg_cm.ravel()
print('\nConfusion Matrix')
print(logreg_cm)
print('True Negative (True Show):\t', true_neg)
print('False Positive (False No-show):\t', false_pos)
print('False Negative (False Show):\t', false_neg)
print('True Positive (True No-show):\t', true_pos)

logreg_cr = classification_report(y_test, y_pred)
print('\nClassification Report')
print(logreg_cr)

# GridSearchCV

# setup hyperparameter grid
c_space = np.logspace(-5, 8, 15)
param_grid = {'C': c_space}

# instantiate GridSearchCV object
logreg_cv = GridSearchCV(logreg, param_grid, cv=5)

# fit data
logreg_cv.fit(X_train, y_train)

# print tuned parameter and score
print('Tuned Logistic Regression Parameters: {}'.format(logreg_cv.best_params_))
print('Best score is {}'.format(logreg_cv.best_score_))

y_pred_cv = logreg_cv.predict(X_test)

# Compute and print the confusion matrix and classification report
logreg_cv_cm = confusion_matrix(y_test, y_pred_cv)
true_pos_cv, false_neg_cv, false_pos_cv, true_neg_cv = logreg_cv_cm.ravel()
print('\nConfusion Matrix (GridSearchCV)')
print(logreg_cv_cm)
print('True Negative (True Show):\t', true_neg_cv)
print('False Positive (False No-show):\t', false_pos_cv)
print('False Negative (False Show):\t', false_neg_cv)
print('True Positive (True No-show):\t', true_pos_cv)

logreg_cv_cr = classification_report(y_test, y_pred_cv)
print('\nClassification Report (GridSearchCV)')
print(logreg_cv_cr)
